refactor: productViewController의 데이터를 다루는 메서드를 별도의 타입으로 분리
